+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - Disallow multiline free external links in the MediaWiki mode
 - Multiline table cells in the MediaWiki mode
 - Allow HTML comments in wikilinks in the MediaWiki mode
+- Allow external links in the image caption in the MediaWiki mode
 
 **Changed**